python3-labgrid: update to version 25.0
This version marks a breaking change. This is the first release using gRPC
instead of crossbar/autobahn for communication between client/exporter and
coordinator.

Version 25.0 exporters / coordinators / clients can not communicate with
Version 24.0 and earlier exporters / coordinators / clients.
All exporters, coordinators and clients in a deployment have to be updated
at the same time.

See the release notes[1] for more information.

This updates the stable recipe as well as the git recipe, since master and
the latest release point to the same commit right now.

[1]: https://github.com/labgrid-project/labgrid/releases/tag/v25.0

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>