python3-labgrid: install helpers
Helpers are not installed automatically, so install them, along with an
(empty) helpers configuration file.

Since helpers are launched with sudo, add it to RDEPENDS as well.

Client/library access to the device via SSH as "root" user is assumed.
Additional configuration (dedicated user and related sudoers config for
password-less helpers usage) shall be done separately.

Signed-off-by: Francesco Valla <francesco@valla.it>